Final Commit for EE201
Hey, just wanted to let you know that the reason why the binding posts are not connect to the circuit is because the vias got clogged, I was in lab for 2 hours today just trying to solder the wires to the circuit but it just was not working. Daniel told me that its okay and I can just submit the photos of my Arduometer like this.